visionML/requirements.txt
ultralytics==8.3.33
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
pyyaml==6.0.1
typer==0.12.5
rich==13.9.4
tqdm==4.66.5
